backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7

# DICOM Processing
pydicom==2.4.4
SimpleITK==2.3.1
nibabel==5.2.0
Pillow==10.1.0
numpy==1.26.2

# AI & ML
torch==2.1.1
transformers==4.35.2
huggingface-hub==0.19.4
openai==1.3.5

# Blockchain
ipfshttpclient==0.8.0a2
web3==6.11.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Storage
boto3==1.29.7

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator>=2.3.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0
